requirements.txt
peewee
PyMySQL
ansicolors
toml
